renamed the packet header/class files, fixed forward declaration. pushing for early use
